requirements.txt
fastapi 
uvicorn 
langchain 
pymupdf 
pillow 
opencv-python 
transformers 
torch
pandas
numpy
python-multipart
python-dotenv
mistralai